添加营销机会_后端

+++ src/main/java/com/gxx/crm/vo/SaleChance.java
@@ -34,7 +34,7 @@ public class SaleChance {
 
     private Integer devResult;
 
-    private Byte isValid;
+    private Integer isValid;
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
     private Date createDate;
